added AwsAccount store, StackInfo object, updated aws account update page, added AWS disclaimer to SC-env files
